requirements.txt
streamlit
requests
python-dotenv
transformers
torch
fpdf
io

